requirements.txt
streamlit>=1.36
pandas>=2.2
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
plotly>=5.22
